fixed parse_immunization_record?
If I understand the comment correctly ("It's possible that multiple vaccines can be given in single day... we should handle this..."), the method should be able to handle multiple immunization records and return a list of them. If this is the case, then I fixed it. If not, then this should be reverted, and a different patch applied.